refactor(jsx,go-template): classify template-literal memo on the IR, drop adapter parse
Move the memo template-literal classification from the Go adapter to the
analyzer. MemoInfo gains bodyIsTemplateLiteral, set from the real arrow AST node
at analysis time; inferMemoType reads the flag instead of re-parsing computation
with ts.createSourceFile, and the isTemplateLiteralMemo helper is removed. A
no-substitution `plain` template folds to a plain string ParsedExpr literal, so
a dedicated boolean (not a parsed.kind check) preserves the backtick
distinction.

Byte-identical: the analyzer logic mirrors the former adapter predicate over the
same source. Verified by go unit (556) + conformance (786). Drops the adapter's
package-wide ts.createSourceFile count from 8 to 7, advancing the "no expression
parsing in adapters" rule by moving classification to Phase 1.
